Ignorerer tombstone på gjennomføringer som har deltakere ref https://github.com/navikt/amt-deltakelser/pull/396 (#397)

+++ amt-aktivitetskort-publisher/src/main/kotlin/no/nav/amt/aktivitetskort/service/KafkaConsumerService.kt
@@ -89,8 +89,15 @@ class KafkaConsumerService(
         value: String?,
     ) {
         if (value == null) {
-            if (id !in deltakerlisteTombstoneBlacklist) {
+            log.info("Mottok tombstone for gjennomføring: $id")
+            val antallDeltakere = deltakerRepository.getAntallDeltakereForDeltakerliste(id)
+            if (id !in deltakerlisteTombstoneBlacklist && antallDeltakere == 0) {
                 deltakerlisteRepository.delete(id)
+            } else {
+                log.error(
+                    "Ignorerer tombstone for $id. " +
+                        "Deltakerliste er svartelistet eller har deltakere. Antall deltakere: $antallDeltakere",
+                )
             }
             return
         }
